requirements.txt
python-dotenv
git+ssh://git@git.seclab.cs.ucsb.edu/cgc/farnsworth#egg=farnsworth
nose>=1.3.7
-nose-timer>=0.5.0
-coverage>=4.0.3
